requirements.txt
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
python-dotenv==1.1.1
openpyxl==3.1.5
mysqlclient>=2.2.7
pymysql==1.1.2

gunicorn>=20.0.0

alembic==1.17.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
et_xmlfile==2.0.0
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.4
pandas==2.3.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
